requirements.txt
numpy
torch
torchvision
pyyaml
tqdm
matplotlib
scikit-learn